Refactor: move parameter handling to vw_base
